requirements.txt
mcp
PyGithub
openai
langchain
langchain-openai
python-dotenv
fastapi
uvicorn